Commented out unneeded code that causes warnings during compile
I might want a double-check on the marine_card.dm change. It doesn't seem to make any difference with regard to access, job assignment on an ID computer, etc, but I think if it did affect the access we would have noticed it a long time ago and fixed the warning.